Genorder class; linkusd trading
Still need to integrate this stuff into the actual application itself.
